requirements.txt
pandas~=2.2.3
apache-airflow~=3.0.2
psycopg2-binary~=2.9.10
pyodbc~=5.2.0
confluent-kafka~=2.10.1
protobuf~=4.25.3
grpcio-tools~=1.59.0
fastavro~=1.11.1
requests~=2.32.3
pyspark~=4.0.0
numpy~=2.1.3
tenacity~=9.1.2